prisoner_dilemma_annotation/requirements.txt
flask==2.0.1
gunicorn==20.1.0
supabase==1.0.3
python-dotenv==0.19.0 